fix seeds, debug titles
