+++ index.html
@@ -83,10 +83,10 @@
 
 </body>
 <!-- App JS all the func.js files should go here, and use doc.ready at the top of pages.-->
-<script src="/scripts/renderStory.js"></script>
-<script src="/scripts/renderTitles.js"></script>
-<script src="/scripts/newContribution.js"></script>
-<script src="/scripts/createStoryElement.js"></script>
-<script src="/scripts/createContributionElement.js"></script>
+<script src="/renderStory.js"></script>
+<script src="/renderTitles.js"></script>
+<script src="/newContribution.js"></script>
+<script src="/createStoryElement.js"></script>
+<script src="/createContributionElement.js"></script>
 
 </html>
