+++ public/scripts/app.js
@@ -10,13 +10,12 @@ $(() => {
 
   //this will make it client side SPA behaviour
   const loadTitles = function() {
-    $.get("/titles", (response) => {
+    $.get("/api/titles", (response) => {
       renderTitles(response);
     });
   };
   loadTitles()
-  // $.get(/*route here*/)
-  //   .then((/*what the route returns*/) => {})
+
 
   // $.post(/*route here*/)
   //   .then((/*what the route returns*/) => {})
