+++ views/index.ejs
@@ -96,5 +96,12 @@
 <script src="/scripts/newContribution.js"></script>
 <script src="/scripts/createStoryElement.js"></script>
 <script src="/scripts/createContributionElement.js"></script>
+<script src="/scripts/error.js"></script>
+<script src="/scripts/onClickTitle.js"></script>
+<script src="/scripts/createTitleElement.js"></script>
+<script src="/scripts/createStorySubmit.js"></script>
+<script src="/scripts/createStoryButton.js"></script>
+<script src="/scripts/contributionSubmit.js"></script>
+<script src="/scripts/app.js"></script>
 
 </html>
